# Conflicts:
#   Backend/Malitell/src/main/java/com/ssafy/malitell/controller/UserController.java

+++ Backend/Malitell/src/main/java/com/ssafy/malitell/repository/UserRepository.java
@@ -3,10 +3,8 @@ package com.ssafy.malitell.repository;
 import com.ssafy.malitell.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Long>, UserRepositoryCustom {
+public interface UserRepository extends JpaRepository<User, Integer>, UserRepositoryCustom {
     Boolean existsByUserId(String userId);
     User findByUserId(String userId);
-
     User findByUserSeq(int userSeq);
-
 }
